tasks/qualification/app/requirements.txt
gunicorn
flask
Pillow
